requirements.txt
aiohttp==3.7.3
aiohttp-cors==0.7.0
aioredis==1.3.1
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
blessings==1.7
blis==0.7.4
bokeh==2.2.3
cachetools==4.2.1
catalogue==1.0.0
certifi==2020.12.5
chardet==3.0.4
clean-text==0.3.0
cleantext==1.1.3
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colorful==0.5.4
coverage==5.4
cymem==2.0.5
dask==2021.1.1
decorator==4.4.2
distributed==2021.1.1
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
emoji==1.2.0
filelock==3.0.12
fsspec==0.8.5
ftfy==5.8
gensim==3.8.3
google-api-core==1.25.1
google-auth==1.24.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
graphviz==0.16
grpcio==1.35.0
HeapDict==1.0.1
hiredis==1.1.0
html2text==2020.1.16
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==8.0.0a0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
locket==0.2.1
MarkupSafe==1.1.1
modin==0.8.3
msgpack==1.0.2
multidict==5.1.0
murmurhash==1.0.5
nbformat==5.1.2
networkx==2.5
nltk==3.5
numpy==1.20.0
nvidia-ml-py3==7.352.0
opencensus==0.7.12
opencensus-context==0.1.2
packaging==20.9
pandas==1.1.5
parso==0.8.1
partd==1.1.0
pathy==0.3.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pkg-resources==0.0.0
plac==1.1.3
preshed==3.0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
py-spy==0.3.4
pyarrow==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
pytextrank==2.1.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.2
ray==1.1.0
redis==3.5.3
regex==2020.11.13
requests==2.25.1
rsa==4.7
scipy==1.6.0
six==1.15.0
smart-open==3.0.0
sortedcontainers==2.3.0
soupsieve==2.1
spacy==2.3.5
spacy-legacy==3.0.1
srsly==1.0.5
swifter==1.0.7
tblib==1.7.0
thinc==7.4.5
toolz==0.11.1
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.7.4.3
unicode==2.8
Unidecode==1.1.2
uritools==3.0.0
urlextract==1.2.0
urllib3==1.26.3
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.0a0
yarl==1.6.3
zict==2.0.0
